Initial setup proof of concept
this verion is working with auth route working and a couple temp routes
setup, to show proof of concept.  I will work on converting more of the
routes.
