+++ API/things-api/routes/a/index.js
@@ -1,0 +1,4 @@
+module.exports = function(app){
+  var stuff = 'something.'
+
+}
