+++ API/things-api/routes/test.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+  res.status(200).json({message: 'You have connected to /test' });
+};
